+++ mongosql/crud/__init__.py
@@ -1,2 +1,3 @@
 from .crudhelper import CrudHelper, StrictCrudHelper
 from .crudview import CrudViewMixin, CRUD_METHOD, saves_relations, ABSENT
+from ..util import EntityDictWrapper, load_many_instance_dicts, model_primary_key_columns_and_names