+++ mongosql/util/__init__.py
@@ -5,3 +5,7 @@ from .reusable import Reusable
 from .mongoquery_settings_handler import MongoQuerySettingsHandler
 from .marker import Marker
 from .settings_dict import MongoQuerySettingsDict, StrictCrudHelperSettingsDict
+from .method_decorator import method_decorator, method_decorator_meta
+from .bulk import \
+    EntityDictWrapper, load_many_instance_dicts, \
+    model_primary_key_columns_and_names, entity_dict_has_primary_key